修复数据类型转换和测试用例
更新了 Valid 类中的数据类型转换方法，确保所有类型的 ToBoolean、ToChar、ToSByte、ToByte、ToInt16、ToUInt16、ToInt32、ToUInt32、ToInt64、ToUInt64、ToSingle、ToDouble、ToDecimal、ToDateTime 和 ToString 方法的实现符合预期。

同时，更新了 DelimiterBasedStringConverterTests 测试用例，确保序列化和反序列化的结果正确，特别是布尔值的表示方式从 "1" 改为 "True"。
