--- src/LuYao.Common/Valid/Valid.Byte.cs
@@ -1,37 +1,0 @@
-﻿using System;
-
-namespace LuYao;
-
-partial class Valid
-{
-    /// <inheritdoc/>
-    public static Boolean ToBoolean(Byte value) => value != 0;
-    /// <inheritdoc/>
-    public static Char ToChar(Byte value) => (Char)value;
-    /// <inheritdoc/>
-    public static SByte ToSByte(Byte value) => Convert.ToSByte(value);
-    /// <inheritdoc/>
-    public static Byte ToByte(Byte value) => value;
-    /// <inheritdoc/>
-    public static Int16 ToInt16(Byte value) => Convert.ToInt16(value);
-    /// <inheritdoc/>
-    public static UInt16 ToUInt16(Byte value) => Convert.ToUInt16(value);
-    /// <inheritdoc/>
-    public static Int32 ToInt32(Byte value) => Convert.ToInt32(value);
-    /// <inheritdoc/>
-    public static UInt32 ToUInt32(Byte value) => Convert.ToUInt32(value);
-    /// <inheritdoc/>
-    public static Int64 ToInt64(Byte value) => Convert.ToInt64(value);
-    /// <inheritdoc/>
-    public static UInt64 ToUInt64(Byte value) => Convert.ToUInt64(value);
-    /// <inheritdoc/>
-    public static Single ToSingle(Byte value) => Convert.ToSingle(value);
-    /// <inheritdoc/>
-    public static Double ToDouble(Byte value) => Convert.ToDouble(value);
-    /// <inheritdoc/>
-    public static Decimal ToDecimal(Byte value) => Convert.ToDecimal(value);
-    /// <inheritdoc/>
-    public static DateTime ToDateTime(Byte value) => default;
-    /// <inheritdoc/>
-    public static String ToString(Byte value) => Convert.ToString(value);
-}
