+++ src/LuYao.Common/Valid/Valid.DateTime.cs
@@ -24,53 +24,6 @@ partial class Valid
         }
     }
 
-    /// <summary>
-    /// 将值转换为字符串。
-    /// </summary>
-    public static string ToString(DateTime? value, Date format) => ToString(value ?? default, format);
-
     /// <inheritdoc/>
     public static string ToString(DateTime value) => value.ToString("O");
-
-    /// <inheritdoc/>
-    public static bool ToBoolean(DateTime value) => false;
-
-    /// <inheritdoc/>
-    public static char ToChar(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static sbyte ToSByte(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static byte ToByte(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static short ToInt16(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static ushort ToUInt16(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static int ToInt32(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static uint ToUInt32(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static long ToInt64(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static ulong ToUInt64(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static float ToSingle(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static double ToDouble(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static decimal ToDecimal(DateTime value) => default;
-
-    /// <inheritdoc/>
-    public static DateTime ToDateTime(DateTime value) => value;
 }
